requirements.txt
-e .

google-cloud-secret-manager==1.0.0
google-cloud-storage>=1.19.1
pymsteams==0.1.13
PyYAML==5.1.2
sendgrid==6.2.1
taosdevopsutils==1.3.1
python-harvest-apiv2==1.0.2